It's hard to believe it but I've just made QuteNote buildable with Visual Studio 2013. It spotted some places with signed/unsigned int implicit type conversion which I've fixed but other than that it was very-very unpleasant experience. I hope I won't need to get into something like that anymore.

+++ core/src/client/Utility.h
@@ -10,7 +10,7 @@ namespace qute_note {
 template <class T>
 bool CheckGuid(const T & guid)
 {
-    size_t guidSize = guid.size();
+    qint32 guidSize = static_cast<qint32>(guid.size());
 
     if (guidSize < qevercloud::EDAM_GUID_LEN_MIN) {
         return false;
